# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SoW API
  description: >-
    Business API for SoW. All operations live under the /v1 prefix.
    Operational probes (/livez, /readyz) and /metrics are infrastructure
    endpoints outside this contract — see
    specs/001-auth-landing-dashboard/contracts/health-and-ops.md.
    Timestamps are Unix milliseconds (int64) everywhere. Session endpoints
    under /auth are documented in
    specs/001-auth-landing-dashboard/contracts/session-cookie.md and
    specs/004-first-party-auth-forms/contracts/.
  version: 0.3.0
servers:
  - url: /v1
    description: Versioned API root
paths:
...
  /auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Complete sign-in or sign-up with the emailed code
      description: >-
        Exchanges the pending token and the six-digit code with the
        identity provider. On success seals the session cookie and returns
        an access token. A wrong, expired, or unknown code answers 400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          $ref: '#/components/responses/SessionStarted'
        '400':
          description: Malformed input (validation_failed) or a rejected code (invalid_code).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/ForbiddenOrigin'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ProviderUnavailable'
  /auth/resend-verification:
    post:
      operationId: resendVerification
      summary: Email a new verification code
      description: >-
        Asks the identity provider to send a fresh code to the address.
        The previous code stops working. Always answers 202 for a
        well-formed address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '202':
          description: Accepted.
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/ForbiddenOrigin'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      summary: Email a password reset link
      description: >-
        Issues a single-use, expiring reset token at the identity provider
        and emails a link to the web app's reset page. Always answers 202
        for a well-formed address.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailRequest'
      responses:
        '202':
          description: Accepted.
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '403':
          $ref: '#/components/responses/ForbiddenOrigin'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/reset-password:
    post:
      operationId: resetPassword
      summary: Set a new password from a reset link
      description: >-
        Confirms the reset token with the identity provider, sets the new
        password, signs the user in, seals the session cookie, and returns
        an access token. An expired or used token answers 400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          $ref: '#/components/responses/SessionStarted'
        '400':
          description: Malformed input (validation_failed) or a rejected token (invalid_reset_token).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          $ref: '#/components/responses/ForbiddenOrigin'
        '422':
          $ref: '#/components/responses/WeakPassword'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ProviderUnavailable'
  /auth/callback:
    get:
      operationId: completeLogin
      summary: Complete hosted sign-in
      description: >-
        Return leg of the hosted flow. Verifies the state, exchanges the
        authorization code with the identity provider server-side,
        provisions the user record, seals the refresh token into the
        session cookie, and redirects to the web app. Every failure
        redirects to the web app's sign-in page with an error code and no
        session.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: >-
            Redirect to WEB_APP_URL/callback?returnTo=… on success, or to
            WEB_APP_URL/?error=… on failure.
          headers:
            Location:
              schema:
                type: string
            Set-Cookie:
              schema:
                type: string
              description: The sealed session cookie, or a clearing cookie on failure.
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: WorkOS AuthKit access token, validated against the WorkOS JWKS.
    sessionCookie:
      type: apiKey
      in: cookie
      name: sow_session
      description: >-
        Encrypted HttpOnly session cookie set by /auth/callback. Read only
        by /auth/refresh and /auth/logout.
  responses:
...
    ForbiddenOrigin:
      description: The Origin header is missing or not an allowed web origin.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ProviderUnavailable:
      description: The identity provider could not be reached.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    SessionStarted:
      description: Signed in. The session cookie is set alongside.
      headers:
        Set-Cookie:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AccessToken'
...
    ValidationFailed:
      description: Malformed input. details lists field errors.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    WeakPassword:
      description: The identity provider rejected the password. details.message says why.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    AccessToken:
      type: object
      required: [access_token, expires_at]
      properties:
        access_token:
          type: string
          description: Short-lived bearer token for /v1 calls. Hold in memory only.
        expires_at:
          type: integer
          format: int64
          description: Token expiry, Unix milliseconds.
...
    VerifyEmailRequest:
      type: object
      required: [pending_token, code]
      properties:
        pending_token:
          type: string
          maxLength: 512
        code:
          type: string
          pattern: '^[0-9]{6}$'
    EmailRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          maxLength: 512
        password:
          type: string
          minLength: 8
          maxLength: 256
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable snake_case error code.
        request_id:
          type: string
          description: Correlation id to quote in a support request.
        details:
          description: Optional structured detail (validation errors and similar).